+++ Puzzles.Runner/2019/Day05.cs
@@ -8,7 +8,7 @@ public class Day05(ILinesInputReader input) : IPuzzleSolver
     private IntCodeMachine _machine;
 
     public void Init()
-        => _machine = new([.. input.GetTokens(",", Convert.ToInt32).First()]);
+        => _machine = new([.. input.GetTokens(",", Convert.ToInt64).First()]);
 
     public string SolvePart1()
     {
